+++ tests/UnitTests/POData/ObjectModel/Serialisers/SerialiserWriteElementTest.php
@@ -9,6 +9,7 @@ use POData\ObjectModel\CynicSerialiser as IronicSerialiser;
 use POData\ObjectModel\ODataBagContent;
 use POData\ObjectModel\ODataCategory;
 use POData\ObjectModel\ODataEntry;
+use POData\ObjectModel\ODataFeed;
 use POData\ObjectModel\ODataLink;
 use POData\ObjectModel\ODataMediaLink;
 use POData\ObjectModel\ODataProperty;
@@ -61,6 +62,7 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $link->type = 'application/atom+xml;type=feed';
         $link->title = 'Orders';
         $link->url = 'Customers(CustomerID=\'1\',CustomerGuid=guid\'123e4567-e89b-12d3-a456-426655440000\')/Orders';
+        $link->isCollection = true;
 
         $propContent = new ODataPropertyContent();
         $propContent->properties = ['CustomerID' => new ODataProperty(), 'CustomerGuid' => new ODataProperty(),
@@ -225,6 +227,17 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $linkPropContent->properties['Address']->typeName = 'Address';
         $linkPropContent->properties['Address']->value = $addressContent;
 
+        $linkRawResult = new ODataFeed();
+        $linkRawResult->id = 'http://localhost/odata.svc/Customers(CustomerID=\'1\','
+                             .'CustomerGuid=guid\'123e4567-e89b-12d3-a456-426655440000\')/Orders';
+        $linkRawResult->title = new ODataTitle('Orders');
+        $linkRawResult->selfLink = new ODataLink();
+        $linkRawResult->selfLink->name = 'self';
+        $linkRawResult->selfLink->title = 'Orders';
+        $linkRawResult->selfLink->url = 'Customers(CustomerID=\'1\','
+                                        .'CustomerGuid=guid\'123e4567-e89b-12d3-a456-426655440000\')/Orders';
+
+
         $linkResult = new ODataEntry();
         $linkResult->id = 'http://localhost/odata.svc/Customers(CustomerID=\'1\',CustomerGuid'
                           .'=guid\'123e4567-e89b-12d3-a456-426655440000\')';
@@ -241,9 +254,20 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $linkResult->links[0]->type = 'application/atom+xml;type=feed';
         $linkResult->links[0]->url = 'Customers(CustomerID=\'1\',CustomerGuid=guid\'123e4567'
                                      .'-e89b-12d3-a456-426655440000\')/Orders';
+        $linkResult->links[0]->isCollection = true;
+        $linkResult->links[0]->isExpanded = true;
+        $linkResult->links[0]->expandedResult = $linkRawResult;
         $linkResult->resourceSetName = 'Customers';
         $linkResult->propertyContent = $linkPropContent;
         $linkResult->updated = '2017-01-01T00:00:00+00:00';
+
+        $linkFeedResult = new ODataFeed();
+        $linkFeedResult->id = 'http://localhost/odata.svc/Orders(OrderID=1)/Order_Details';
+        $linkFeedResult->title = new ODataTitle('Order_Details');
+        $linkFeedResult->selfLink = new ODataLink();
+        $linkFeedResult->selfLink->name = 'self';
+        $linkFeedResult->selfLink->title = 'Order_Details';
+        $linkFeedResult->selfLink->url = 'Orders(OrderID=1)/Order_Details';
 
         $links = [new ODataLink(), new ODataLink()];
         $links[0]->name = 'http://schemas.microsoft.com/ado/2007/08/dataservices/related/Customer';
@@ -257,6 +281,9 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $links[1]->title = 'Order_Details';
         $links[1]->type = 'application/atom+xml;type=feed';
         $links[1]->url = 'Orders(OrderID=1)/Order_Details';
+        $links[1]->isCollection = true;
+        $links[1]->isExpanded = true;
+        $links[1]->expandedResult = $linkFeedResult;
 
         $objectResult = new ODataEntry();
         $objectResult->id = 'http://localhost/odata.svc/Orders(OrderID=1)';
@@ -347,6 +374,8 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $links[1]->title = 'Subordinates';
         $links[1]->type = 'application/atom+xml;type=feed';
         $links[1]->url = 'Employees(EmployeeID=\'Cave+Johnson\')/Subordinates';
+        $links[1]->isCollection = true;
+        $links[1]->isExpanded = false;
 
         $objectResult = new ODataEntry();
         $objectResult->id = 'http://localhost/odata.svc/Employees(EmployeeID=\'Cave+Johnson\')';
@@ -414,15 +443,16 @@ class SerialiserWriteElementTest extends SerialiserTestBase
         $expandNode->shouldReceive('getResourceProperty')->andReturn($rProp);
         $expandNode->shouldReceive('getPropertyName')->andReturn('Manager');
         $expandNode->shouldReceive('canSelectAllProperties')->andReturn(true);
-        $expandNode->shouldReceive('isExpansionSpecified')->andReturn(true);
+        $expandNode->shouldReceive('isExpansionSpecified')->andReturn(false);
         $expandNode->shouldReceive('findNode')->andReturn(null);
 
         $node = m::mock(RootProjectionNode::class);
         $node->shouldReceive('getPropertyName')->andReturn('Customer');
-        $node->shouldReceive('isExpansionSpecified')->andReturn(true, true, true, true);
+        $node->shouldReceive('isExpansionSpecified')->withArgs(['Subordinates'])->andReturn(false)->times(0);
+        $node->shouldReceive('isExpansionSpecified')->withArgs(['Manager'])->andReturn(true)->times(0);
         $node->shouldReceive('canSelectAllProperties')->andReturn(false);
-        $node->shouldReceive('findNode')->andReturn($expandNode);
-        $node->shouldReceive('getChildNodes')->andReturn([$expandNode, $mailNode])->atLeast(1);
+        $node->shouldReceive('findNode')->andReturn($expandNode, $expandNode, null);
+        $node->shouldReceive('getChildNodes')->andReturn([$expandNode, $mailNode])->times(1);
 
         $ironic->getRequest()->setRootProjectionNode($node);
 
@@ -498,16 +528,25 @@ class SerialiserWriteElementTest extends SerialiserTestBase
             ''
         );
 
+        $managerResult = new ODataEntry();
+        $managerResult->resourceSetName = 'Employee';
+
         $managerLink1 = new ODataLink();
         $managerLink1->name = 'http://schemas.microsoft.com/ado/2007/08/dataservices/related/Manager';
         $managerLink1->title = 'Manager';
         $managerLink1->type = 'application/atom+xml;type=entry';
         $managerLink1->url = 'Employees(EmployeeID=\'Cave+Johnson\')/Manager';
+        $managerLink1->isCollection = false;
+        $managerLink1->isExpanded = true;
+        $managerLink1->expandedResult = $managerResult;
         $managerLink2 = new ODataLink();
         $managerLink2->name = 'http://schemas.microsoft.com/ado/2007/08/dataservices/related/Subordinates';
         $managerLink2->title = 'Subordinates';
         $managerLink2->type = 'application/atom+xml;type=feed';
         $managerLink2->url = 'Employees(EmployeeID=\'Cave+Johnson\')/Subordinates';
+        $managerLink2->isCollection = true;
+        $managerLink2->isExpanded = false;
+
 
         $manager = new ODataEntry();
         $manager->id = 'http://localhost/odata.svc/Employees(EmployeeID=\'Cave+Johnson\')';