+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
-   * Return HTTP 204 when a singleton resource is null (see [#205] (https://github.com/Algo-Web/POData/pull/201))
+   * Fix empty-object serialisation - thanks to **cdcampos** for reporting this (see [#206] (https://github.com/Algo-Web/POData/pull/206))
+   * Return HTTP 204 when a singleton resource is null (see [#205] (https://github.com/Algo-Web/POData/pull/205))
    * Fix property expansion - **cdcampos** is on a roll! (see [#201] (https://github.com/Algo-Web/POData/pull/201))
    * Fix ODataCategory and ODataLink writes via JSON - **cdcampos** rides again! (see [#197] (https://github.com/Algo-Web/POData/pull/197))
    * Untangle batch sub-request handling (see [#191] (https://github.com/Algo-Web/POData/pull/191))